requirements.txt
requests
feedparser
beautifulsoup4
openai
python-dotenv